유저 자신의 페르소나만 수정 삭제

+++ fillme/mypage/views.py
@@ -65,7 +65,7 @@ def my_persona_list_create(request):
         personas = user.persona_set.all()
         serializer = PersonaSerializer(personas, many=True)
         return Response(serializer.data)
-
+        
 # 페르소나 조회/수정/삭제
 @api_view(['GET','PATCH', 'DELETE'])
 @permission_classes([IsAuthenticatedOrReadOnly])
@@ -77,18 +77,20 @@ def my_persona_rud(request, persona_id):
         serializer = PersonaSerializer(persona)
         return Response(serializer.data)
     elif request.method == "PATCH":
-        serializer=PersonaSerializer(data={
-            'user':user.id,
-            'profile':profile.id,
-            'name':request.data['name'],
-            'category':request.data['category'],
-            'image':request.data['image']},instance=persona)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-        return Response(serializer.data)
+        if user == persona.user:
+            serializer=PersonaSerializer(data={
+                'user':user.id,
+                'profile':profile.id,
+                'name':request.data['name'],
+                'category':request.data['category'],
+                'image':request.data['image']},instance=persona)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+            return Response(serializer.data)
     elif request.method == "DELETE":
-        persona.delete()
-        return Response({'persona_id':persona_id})
+        if user == persona.user:
+            persona.delete()
+            return Response({'persona_id':persona_id})
 
 # 다른 유저의 페르소나 목록 조회
 @api_view(['GET'])
